feat: interpreter 메서드 구현
- 연산자, 항 두개를 받아 연산결과 반환

+++ src/main/java/calculator/domain/Calculator.java
@@ -1,19 +1,19 @@
 package calculator.domain;
 
 public class Calculator {
-    public static int add(int num1, int num2) {
+    public static long add(long num1, long num2) {
         return num1 + num2;
     }
 
-    public static int sub(int num1, int num2) {
+    public static long sub(long num1, long num2) {
         return num1 - num2;
     }
 
-    public static int mul(int num1, int num2) {
+    public static long mul(long num1, long num2) {
         return num1 * num2;
     }
 
-    public static int div(int num1, int num2) {
+    public static long div(long num1, long num2) {
         return num1 / num2;
     }
 }
