+++ src/test/java/calculator/domain/CalculatorTest.java
@@ -9,25 +9,25 @@ class CalculatorTest {
 
     @Test
     void add() {
-        int result = Calculator.add(1, 3);
+        long result = Calculator.add(1, 3);
         assertThat(result).isEqualTo(4);
     }
 
     @Test
     void sub() {
-        int result = Calculator.sub(3, 1);
+        long result = Calculator.sub(3, 1);
         assertThat(result).isEqualTo(2);
     }
 
     @Test
     void mul() {
-        int result = Calculator.mul(2, 3);
+        long result = Calculator.mul(2, 3);
         assertThat(result).isEqualTo(6);
     }
 
     @Test
     void div() {
-        int result = Calculator.div(6, 3);
+        long result = Calculator.div(6, 3);
         assertThat(result).isEqualTo(2);
     }
